+++ nrf_802154/driver/src/nrf_802154.c
@@ -60,6 +60,7 @@
 #include "nrf_802154_rx_buffer.h"
 #include "nrf_802154_tx_power.h"
 #include "nrf_802154_stats.h"
+#include "nrf_802154_swi.h"
 #include "hal/nrf_radio.h"
 #include "platform/nrf_802154_clock.h"
 #include "platform/nrf_802154_random.h"
@@ -98,6 +99,15 @@ static inline bool are_frame_properties_valid(const nrf_802154_transmitted_frame
 static inline bool are_extra_cca_attempts_valid(const nrf_802154_transmit_at_metadata_t * p_data)
 {
     return !p_data->cca || (p_data->extra_cca_attempts < UINT8_MAX);
+}
+
+static inline bool is_tx_timestamp_request_valid(const bool tx_timestamp_encode)
+{
+#if NRF_802154_TX_TIMESTAMP_PROVIDER_ENABLED
+    return true;
+#else
+    return !tx_timestamp_encode;
+#endif
 }
 
 void nrf_802154_channel_set(uint8_t channel)
@@ -182,6 +192,7 @@ void nrf_802154_init(void)
         .exit  = nrf_802154_critical_section_exit
     };
 
+    nrf_802154_swi_init();
     nrf_802154_ack_data_init();
     nrf_802154_core_init();
     nrf_802154_clock_init();
@@ -398,10 +409,11 @@ bool nrf_802154_transmit_raw(uint8_t                              * p_data,
     {
         static const nrf_802154_transmit_metadata_t metadata_default =
         {
-            .frame_props = NRF_802154_TRANSMITTED_FRAME_PROPS_DEFAULT_INIT,
-            .cca         = true,
-            .tx_power    = {.use_metadata_value = false},
-            .tx_channel  = {.use_metadata_value = false}
+            .frame_props         = NRF_802154_TRANSMITTED_FRAME_PROPS_DEFAULT_INIT,
+            .cca                 = true,
+            .tx_power            = {.use_metadata_value = false},
+            .tx_channel          = {.use_metadata_value = false},
+            .tx_timestamp_encode = false
         };
 
         p_metadata = &metadata_default;
@@ -413,19 +425,21 @@ bool nrf_802154_transmit_raw(uint8_t                              * p_data,
 
     nrf_802154_transmit_params_t params =
     {
-        .frame_props        = p_metadata->frame_props,
-        .tx_power           = {0},
-        .channel            = channel,
-        .cca                = p_metadata->cca,
-        .immediate          = false,
-        .extra_cca_attempts = 0U,
+        .frame_props         = p_metadata->frame_props,
+        .tx_power            = {0},
+        .channel             = channel,
+        .cca                 = p_metadata->cca,
+        .immediate           = false,
+        .extra_cca_attempts  = 0U,
+        .tx_timestamp_encode = p_metadata->tx_timestamp_encode,
     };
 
     (void)nrf_802154_tx_power_convert_metadata_to_tx_power_split(channel,
                                                                  p_metadata->tx_power,
                                                                  &params.tx_power);
 
-    result = are_frame_properties_valid(&params.frame_props);
+    result = are_frame_properties_valid(&params.frame_props) &&
+             is_tx_timestamp_request_valid(p_metadata->tx_timestamp_encode);
     if (result)
     {
         result = nrf_802154_request_transmit(NRF_802154_TERM_NONE,
@@ -447,10 +461,11 @@ bool nrf_802154_transmit_raw_at(uint8_t                                 * p_data
     bool                              result;
     nrf_802154_transmit_at_metadata_t metadata_default =
     {
-        .frame_props        = NRF_802154_TRANSMITTED_FRAME_PROPS_DEFAULT_INIT,
-        .cca                = true,
-        .tx_power           = {.use_metadata_value = false},
-        .extra_cca_attempts = 0,
+        .frame_props         = NRF_802154_TRANSMITTED_FRAME_PROPS_DEFAULT_INIT,
+        .cca                 = true,
+        .tx_power            = {.use_metadata_value = false},
+        .extra_cca_attempts  = 0,
+        .tx_timestamp_encode = false,
     };
 
     nrf_802154_log_function_enter(NRF_802154_LOG_VERBOSITY_LOW);
@@ -462,7 +477,8 @@ bool nrf_802154_transmit_raw_at(uint8_t                                 * p_data
     }
 
     result = are_frame_properties_valid(&p_metadata->frame_props) &&
-             are_extra_cca_attempts_valid(p_metadata);
+             are_extra_cca_attempts_valid(p_metadata) &&
+             is_tx_timestamp_request_valid(p_metadata->tx_timestamp_encode);
     if (result)
     {
         result = nrf_802154_request_transmit_raw_at(p_data, tx_time, p_metadata);
@@ -726,15 +742,17 @@ bool nrf_802154_transmit_csma_ca_raw(uint8_t                                    
     {
         static const nrf_802154_transmit_csma_ca_metadata_t metadata_default =
         {
-            .frame_props = NRF_802154_TRANSMITTED_FRAME_PROPS_DEFAULT_INIT,
-            .tx_power    = {.use_metadata_value = false},
-            .tx_channel  = {.use_metadata_value = false}
+            .frame_props         = NRF_802154_TRANSMITTED_FRAME_PROPS_DEFAULT_INIT,
+            .tx_power            = {.use_metadata_value = false},
+            .tx_channel          = {.use_metadata_value = false},
+            .tx_timestamp_encode = false
         };
 
         p_metadata = &metadata_default;
     }
 
-    result = are_frame_properties_valid(&p_metadata->frame_props);
+    result = are_frame_properties_valid(&p_metadata->frame_props) &&
+             is_tx_timestamp_request_valid(p_metadata->tx_timestamp_encode);
     if (result)
     {
         result = nrf_802154_request_csma_ca_start(p_data, p_metadata);