+++ nrf_802154/sl/sl_opensource/include/nrf_802154_sl_periphs.h
@@ -18,8 +18,18 @@
 #include "nrf_802154_sl_periphs_internal.h"
 #endif
 
+/**
+ * @def NRF_802154_SL_EGU_CLOCK_CHANNEL_NO
+ *
+ * Number of the EGU channel used for HFCLK control.
+ *
+ */
+#ifndef NRF_802154_SL_EGU_CLOCK_CHANNEL_NO
+#define NRF_802154_SL_EGU_CLOCK_CHANNEL_NO   5
+#endif
+
 #define NRF_802154_SL_PPI_CHANNELS_USED_MASK 0U
 #define NRF_802154_SL_PPI_GROUPS_USED_MASK   0U
-#define NRF_802154_SL_EGU_USED_CHANNELS_MASK 0U
+#define NRF_802154_SL_EGU_USED_CHANNELS_MASK (1U << NRF_802154_SL_EGU_CLOCK_CHANNEL_NO)
 
 #endif // NRF_802154_SL_PERIPHS_H__